+++ source_images/feep.ppm
@@ -1,0 +1,1 @@
+P3 4 4 255 0 0 0 0 0 0 0 0 0 255 0 255 0 0 0 0 255 127 0 0 0 0 0 0 0 0 0 0 0 0 0 255 127 0 0 0 255 0 255 0 0 0 0 0 0 0 0 0
