+++ source_images/tiny.ppm
@@ -1,0 +1,1 @@
+P3 1 1 255 123 234 95
